+++ src/components/map/MapApi.js
@@ -1,13 +1,15 @@
 import MapApi, {
     layerTypes,
     controlTypes,
-    loadEarthEngineApi,
+    loadEarthEngineWorker,
     poleOfInaccessibility,
 } from '@dhis2/maps-gl';
 import getMapLocale from './mapLocale';
 
 // Returns a new map instance
 const map = options => {
+    const baseUrl = process.env.DHIS2_BASE_URL;
+    const glyphs = `${baseUrl}/dhis-web-maps/fonts/{fontstack}/{range}.pbf`;
     const div = document.createElement('div');
 
     div.className = 'dhis2-map';
@@ -17,9 +19,15 @@ const map = options => {
     return new MapApi(div, {
         ...options,
         locale: getMapLocale(),
+        glyphs,
     });
 };
 
-export { layerTypes, controlTypes, loadEarthEngineApi, poleOfInaccessibility };
+export {
+    layerTypes,
+    controlTypes,
+    loadEarthEngineWorker,
+    poleOfInaccessibility,
+};
 
 export default map;