+++ src/components/map/MapLoadingMask.js
@@ -3,7 +3,7 @@ import { ComponentCover, CenteredContent, CircularLoader } from '@dhis2/ui';
 import styles from './styles/MapLoadingMask.module.css';
 
 const MapLoadingMask = () => (
-    <ComponentCover translucent className={styles.cover}>
+    <ComponentCover className={styles.cover}>
         <CenteredContent>
             <CircularLoader />
         </CenteredContent>